Re-organise code somewhat
