+++ .gitignore
@@ -36,3 +36,4 @@ hs_err*.log
 *~.nib
 *thumbs.db
 bin/
+env
