+++ src/main/java/qupath/bioimageio/spec/BioimageIoSpec.java
@@ -9,12 +9,9 @@ import java.net.URI;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -23,7 +20,6 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-import com.google.gson.JsonArray;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.yaml.snakeyaml.LoaderOptions;
@@ -31,10 +27,7 @@ import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.constructor.SafeConstructor;
 
 import com.google.gson.GsonBuilder;
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 import com.google.gson.annotations.SerializedName;
 import com.google.gson.reflect.TypeToken;
@@ -59,407 +52,6 @@ import com.google.gson.reflect.TypeToken;
 public class BioimageIoSpec {
 	
 	private final static Logger logger = LoggerFactory.getLogger(BioimageIoSpec.class);
-
-	/**
-	 * Parse a model from a file or directory.
-	 * This can either be a yaml file, or a directory that contains a yaml file representing the model.
-	 * @param file The file containing the YAML, or its parent directory.
-	 * @return The parsed model.
-	 * @throws IOException if the model cannot be found or parsed
-	 */
-	public static BioimageIoModel parseModel(File file) throws IOException {
-		return parseModel(file.toPath());
-	}
-
-	/**
-	 * Parse a model from a path.
-	 * This can either represent a yaml file, or a directory that contains a yaml file representing the model.
-	 * @param path The path to the file containing the YAML, or its parent directory.
-	 * @return The parsed model.
-	 * @throws IOException if the model cannot be found or parsed
-	 */
-	public static BioimageIoModel parseModel(Path path) throws IOException {
-		
-		var pathYaml = findModelRdf(path);
-		if (pathYaml == null) {
-			throw new IOException("Can't find rdf.yaml from " + path);
-		}
-		
-		try (var stream = Files.newInputStream(pathYaml)) {
-			var model = parseModel(stream);
-			if (model != null) {
-				model.baseURI = pathYaml.getParent().toUri();
-				model.uri = pathYaml.toUri();
-			}
-			return model;
-		}
-	}
-	
-	/**
-	 * Parse a model from an input stream.
-	 * Note that {@link BioimageIoModel#getBaseURI()} will return null in this case, because the base URI 
-	 * is unknown.
-	 * @param stream A stream of YAML.
-	 * @return The parsed model.
-	 * @throws IOException if the model cannot be found or parsed
-	 */
-	public static BioimageIoModel parseModel(InputStream stream) throws IOException {
-		
-		try {
-			// Important to use SafeConstructor to restrict potential classes that might be initiated
-			// Note: we can't support -inf or inf (but can support -.inf or .inf)
-			Yaml yaml = new Yaml(new SafeConstructor(new LoaderOptions()));
-			Map<String, ?> map = yaml.load(stream);
-
-			var builder = new GsonBuilder()
-					.serializeSpecialFloatingPointValues()
-					.setPrettyPrinting()
-					.registerTypeAdapter(BioimageIoModel.class, new ModelDeserializer())
-					.registerTypeAdapter(BioimageIoResource.class, new ResourceDeserializer())
-					.registerTypeAdapter(BioimageIoDataset.class, new DatasetDeserializer())
-					.registerTypeAdapter(double[].class, new DoubleArrayDeserializer())
-					.registerTypeAdapter(Author.class, new Author.Deserializer())
-					.registerTypeAdapter(WeightsEntry.class, new WeightsEntry.Deserializer())
-					.registerTypeAdapter(WeightsMap.class, new WeightsMap.Deserializer())
-					.registerTypeAdapter(Shape.class, new ShapeDeserializer())
-					.registerTypeAdapter(Processing.class, new Processing.ProcessingDeserializer())
-					.registerTypeAdapter(Axis[].class, new AxesDeserializer())
-					.registerTypeAdapter(Processing.ProcessingMode.class, new Processing.ModeDeserializer())
-					.registerTypeAdapter(TensorDataDescription.class, new TensorDataDescriptionDeserializer())
-					.setDateFormat("yyyy-MM-dd'T'HH:mm:ss");
-
-
-			var gson = builder.create();
-
-			var json = gson.toJson(map);
-			
-			return gson.fromJson(json, BioimageIoModel.class);
-		} catch (Exception e) {
-            throw new IOException(e);
-		}
-	}
-
-
-
-	/**
-	 * Create a shape array for a given axes.
-	 * The axes are expected to a string containing only the characters
-	 * {@code bitczyx} as defined in the spec.
-	 * <p>
-	 * The purpose of this is to build shape arrays easily without needing to 
-	 * explicitly handle different axes and dimension ordering.
-	 * <p>
-	 * An example:
-	 * <pre>
-	 * <code>
-	 * int[] shape = createShapeArray("byxc", Map.of('x', 256, 'y', 512), 1);
-	 * </code>
-	 * </pre>
-	 * <p>
-	 * This should result in an int array with values {@code [1, 512, 256, 1]}.
-	 * 
-	 * @param axes the axes string
-	 * @param target map defining the intended length for specified dimensions
-	 * @param defaultLength the default length to use for any dimension that are not included in the target map
-	 * @return an int array with the same length as the axes string, containing the requested dimensions or default values
-	 */
-	public static int[] createShapeArray(String axes, Map<Character, Integer> target, int defaultLength) {
-		int[] array = new int[axes.length()];
-		int i = 0;
-		for (var c : axes.toLowerCase().toCharArray()) {
-			array[i] = target.getOrDefault(c, defaultLength);
-			i++;
-		}
-		return array;
-	}
-	
-	/**
-	 * List of file names that may contain the model.
-	 * Names should be checked in order, with preference given to the first that is found.
-	 */
-	static final List<String> MODEL_NAMES = List.of("model.yaml", "model.yml", "rdf.yaml", "rdf.yml");
-	
-	static Path findModelRdf(Path path) throws IOException {
-		return findRdf(path, MODEL_NAMES);
-	}
-	
-	private static Path findRdf(Path path, Collection<String> names) throws IOException {
-		if (isYamlPath(path)) {
-			if (names.isEmpty())
-				return path;
-			else {
-				var name = path.getFileName().toString().toLowerCase();
-				if (names.contains(name) || name.startsWith("model") || name.startsWith("rdf"))
-					return path;
-				return null;
-			}
-		}
-		
-		
-		if (Files.isDirectory(path)) {
-			// Check directory
-			try (Stream<Path> pathStream = Files.list(path)) {
-				List<Path> yamlFiles = pathStream.filter(BioimageIoSpec::isYamlPath).collect(Collectors.toList());
-				if (yamlFiles.isEmpty())
-					return null;
-				if (yamlFiles.size() == 1)
-					return yamlFiles.get(0);
-				for (var name : MODEL_NAMES) {
-					var modelFile = yamlFiles.stream()
-							.filter(p -> p.getFileName().toString().equalsIgnoreCase(name))
-							.findFirst()
-							.orElse(null);
-					if (modelFile != null)
-						return modelFile;
-				}
-			}
-		} else if (path.toAbsolutePath().toString().toLowerCase().endsWith(".zip")) {
-			// Check zip file
-			try (var fs = FileSystems.newFileSystem(path, (ClassLoader)null)) {
-				for (var dir : fs.getRootDirectories()) {
-					for (var name : MODEL_NAMES) {
-						var p = dir.resolve(name);
-						if (Files.exists(p))
-							return p;
-					}
-				}
-			}
-		}
-		return null;
-	}
-	
-	static boolean isYamlPath(Path path) {
-		if (Files.isRegularFile(path)) {
-			var name = path.getFileName().toString().toLowerCase();
-			return name.endsWith(".yaml") || name.endsWith(".yml");
-		}
-		return false;
-	}
-
-	private static class ResourceDeserializer implements JsonDeserializer<BioimageIoResource> {
-		
-		@Override
-		public BioimageIoResource deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) {
-			if (json.isJsonNull())
-				return null;
-			var obj = json.getAsJsonObject();
-			BioimageIoResource resource = new BioimageIoResource();
-			deserializeResourceFields(resource, obj, context, true);
-			return resource;
-		}
-		
-	}
-	
-	private static class DatasetDeserializer implements JsonDeserializer<BioimageIoDataset> {
-		
-		@Override
-		public BioimageIoDataset deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) {
-			if (json.isJsonNull())
-				return null;
-			var obj = json.getAsJsonObject();
-			BioimageIoDataset dataset = new BioimageIoDataset();
-			// Deserialize, but not strictly... i.e. allow nulls because we might just have an ID
-			deserializeResourceFields(dataset, obj, context, false);
-			return dataset;
-		}
-		
-	}
-	
-	
-	private static void deserializeResourceFields(BioimageIoResource resource, JsonObject obj, JsonDeserializationContext context, boolean doStrict) {
-		
-		resource.formatVersion = deserializeField(context, obj, "format_version", String.class, doStrict);
-		resource.authors = deserializeField(context, obj, "authors", parameterizedListType(Author.class), doStrict);
-		resource.description = deserializeField(context, obj, "description", String.class, doStrict);
-		resource.documentation = deserializeField(context, obj, "documentation", String.class, doStrict);
-		
-		resource.license = deserializeField(context, obj, "license", String.class, doStrict);
-		resource.name = deserializeField(context, obj, "name", String.class, doStrict);
-		
-		if (obj.has("attachments")) {
-			if (obj.get("attachments").isJsonObject())
-				resource.attachments = deserializeField(context, obj, "attachments", Map.class, Collections.emptyMap());
-			else
-				logger.warn("Can't parse attachments (not a dict)");
-		}
-		resource.badges = deserializeField(context, obj, "badges", parameterizedListType(Badge.class), Collections.emptyList());
-		resource.cite = deserializeField(context, obj, "cite", parameterizedListType(CiteEntry.class), Collections.emptyList());
-
-		resource.covers = deserializeField(context, obj, "covers", parameterizedListType(String.class), Collections.emptyList());
-		resource.downloadURL = deserializeField(context, obj, "download_url", String.class, null);
-		resource.gitRepo = deserializeField(context, obj, "git_repo", String.class, null);			
-		resource.icon = deserializeField(context, obj, "icon", String.class, null);
-		resource.id = deserializeField(context, obj, "id", String.class, null);
-		resource.links = deserializeField(context, obj, "links", parameterizedListType(String.class), Collections.emptyList());
-		resource.maintainers = deserializeField(context, obj, "maintainers", parameterizedListType(Author.class), Collections.emptyList());
-
-		resource.rdfSource = deserializeField(context, obj, "rdf_source", String.class, null);
-
-		resource.tags = deserializeField(context, obj, "tags", parameterizedListType(String.class), Collections.emptyList());
-
-		resource.trainingData = deserializeField(context, obj, "training_data", BioimageIoDataset.class, null);
-
-		resource.version = deserializeField(context, obj, "version", String.class, null);
-	}
-	
-	
-	
-	private static void deserializeModelFields(BioimageIoModel model, JsonObject obj, JsonDeserializationContext context, boolean doStrict) {
-		model.inputs = deserializeField(context, obj, "inputs", parameterizedListType(InputTensor.class), doStrict);
-		List<BaseTensor> tensors = new ArrayList<>(List.copyOf(model.inputs));
-		for (var tensor: model.inputs) {
-			tensor.validate(tensors);
-		}
-
-
-		if (model.isNewerThan("0.5.0")) {
-			model.testInputs = List.of();
-			model.testOutputs = List.of();
-			model.timestamp = "";
-		} else {
-			// now part of the tensor spec:
-			model.testInputs = deserializeField(context, obj, "test_inputs", parameterizedListType(String.class), doStrict);
-			// now part of the tensor spec:
-			model.testOutputs = deserializeField(context, obj, "test_outputs", parameterizedListType(String.class), doStrict);
-			// removed...?
-			model.timestamp = deserializeField(context, obj, "timestamp", String.class, doStrict);
-		}
-
-		model.weights = deserializeField(context, obj, "weights", WeightsMap.class, doStrict);
-
-		model.config = deserializeField(context, obj, "config", Map.class, Collections.emptyMap());
-
-		model.outputs = deserializeField(context, obj, "outputs", parameterizedListType(OutputTensor.class), Collections.emptyList());
-		tensors.addAll(model.outputs);
-		for (var tensor: model.outputs) {
-			tensor.validate(tensors);
-		}
-		model.packagedBy = deserializeField(context, obj, "packaged_by", parameterizedListType(Author.class), Collections.emptyList());
-
-		model.parent = deserializeField(context, obj, "parent", ModelParent.class, null);
-
-		if (obj.has("run_mode")) {
-			if (obj.get("run_mode").isJsonObject())
-				model.runMode = deserializeField(context, obj, "run_mode", Map.class, Collections.emptyMap());
-			else
-				logger.warn("Can't parse run_mode (not an object)");
-		}
-
-		model.sampleInputs = deserializeField(context, obj, "sample_inputs", parameterizedListType(String.class), Collections.emptyList());
-		model.sampleOutputs = deserializeField(context, obj, "sample_outputs", parameterizedListType(String.class), Collections.emptyList());
-	}
-	
-	private static Type parameterizedListType(Type typeOfList) {
-		return TypeToken.getParameterized(List.class, typeOfList).getType();
-	}
-	
-	private static class ModelDeserializer implements JsonDeserializer<BioimageIoModel> {
-		
-		@Override
-		public BioimageIoModel deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-				throws JsonParseException {
-			
-			if (json.isJsonNull())
-				return null;
-			
-			var model = new BioimageIoModel();
-			
-			var obj = json.getAsJsonObject();
-			
-			deserializeResourceFields(model, obj, context, true);
-			deserializeModelFields(model, obj, context, true);
-
-			return model;
-		}
-	}
-
-	enum TensorType {
-		INPUT,
-		OUTPUT
-	}
-	
-	private static class TensorDeserializer implements JsonDeserializer<BaseTensor> {
-		private final TensorType type;
-
-		TensorDeserializer(TensorType type) {
-			this.type = type;
-		}
-
-		@Override
-		public BaseTensor deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
-			return null;
-		}
-	}
-	
-	private static class ShapeDeserializer implements JsonDeserializer<Shape> {
-
-		@Override
-		public Shape deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-				throws JsonParseException {
-			
-			if (json.isJsonNull())
-				return null;
-			
-			if (json.isJsonArray()) {
-				var shape = new Shape();
-				shape.shape = context.deserialize(json, int[].class);
-				return shape;
-			}
-			var obj = json.getAsJsonObject();
-			if (obj.has("min") && obj.has("step")) {
-				return context.deserialize(obj, Shape.ParameterizedInputShape.class);
-			}
-			if (obj.has("offset") && obj.has("scale")) {
-				return context.deserialize(obj, Shape.ImplicitOutputShape.class);
-			}
-			throw new JsonParseException("Can't deserialize unknown shape: " + json);
-		}
-		
-	}
-	
-	
-	/**
-	 * Deserialize a field from a JSON object.
-	 * @param <T> The type of the field.
-	 * @param context The context used for deserialization.
-	 * @param obj The JSON object that contains the field.
-	 * @param name The name of the field.
-	 * @param typeOfT The type of the field.
-	 * @param doStrict if true, fail if the field is missing; otherwise, return null
-	 * @return A parsed T object.
-	 * @throws IllegalArgumentException if doStrict is true and the field is not found
-	 */
-	private static <T> T deserializeField(JsonDeserializationContext context, JsonObject obj, String name, Type typeOfT, boolean doStrict) throws IllegalArgumentException {
-		if (doStrict && !obj.has(name))
-			throw new IllegalArgumentException("Required field " + name + " not found");
-		return deserializeField(context, obj, name, typeOfT, null);
-	}
-
-	private static <T> T deserializeField(JsonDeserializationContext context, JsonObject obj, String name, Type typeOfT, T defaultValue) {
-		if (obj.has(name)) {
-			return ensureUnmodifiable(context.deserialize(obj.get(name), typeOfT));
-		}
-		return ensureUnmodifiable(defaultValue);
-	}
-	
-	/**
-	 * Minor optimization - ensure any lists, maps or sets are unmodifiable at this point, 
-	 * to avoid generating new unmodifiable wrappers later.
-	 * @param <T> The type of the object.
-	 * @param input A collection that should be made unmodifiable (copied if not already unmodifiable).
-	 * @return An unmodifiable object of class T.
-	 */
-	@SuppressWarnings("unchecked")
-	private static <T> T ensureUnmodifiable(T input) {
-		if (input instanceof List)
-			return (T)Collections.unmodifiableList((List<?>)input);
-		if (input instanceof Map)
-			return (T)Collections.unmodifiableMap((Map<?, ?>)input);		
-		if (input instanceof Set)
-			return (T)Collections.unmodifiableSet((Set<?>)input);		
-		return input;
-	}
-	
 	
 	/**
 	 * General resource, based upon the RDF.
@@ -468,40 +60,40 @@ public class BioimageIoSpec {
 	 */
 	public static class BioimageIoResource {
 				
-		private String formatVersion;
-		private List<Author> authors;
-		private String description;
-		private String documentation;
+		String formatVersion;
+		List<Author> authors;
+		String description;
+		String documentation;
 
-		private String name;
+		String name;
 		
-		private List<String> tags;
+		List<String> tags;
 		
 		@SerializedName("training_data")
-		private BioimageIoDataset trainingData;
+        BioimageIoDataset trainingData;
 
-		private String version;
+		String version;
 		
-		private Map<String,?> attachments;
-		private List<Badge> badges;
-		private List<CiteEntry> cite;
+		Map<String,?> attachments;
+		List<Badge> badges;
+		List<CiteEntry> cite;
 		
-		private List<String> covers;
+		List<String> covers;
 
 		@SerializedName("download_url")
-		private String downloadURL;
-		private String gitRepo;
-		private String icon;
-		private String id;
-		private String license;
+        String downloadURL;
+		String gitRepo;
+		String icon;
+		String id;
+		String license;
 
-		private List<String> links;
-		private List<Author> maintainers;
+		List<String> links;
+		List<Author> maintainers;
 		
 		private String source;
 		
 		@SerializedName("rdf_source")
-		private String rdfSource;
+        String rdfSource;
 
 		public String getVersion() {
 			return version;
@@ -619,61 +211,27 @@ public class BioimageIoSpec {
 			return key;
 		}
 		
-		private static WeightsEntry fromKey(String name) {
+		static WeightsEntry fromKey(String name) {
 			for (var v : values()) {
 				if (v.key.equals(name) || v.alternatives.contains(name))
 					return v;
 			}
 			return null;
 		}
-		
-		
-		private static class Deserializer implements JsonDeserializer<WeightsEntry> {
-
-			@Override
-			public WeightsEntry deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-					throws JsonParseException {
-				return WeightsEntry.fromKey(json.toString());
-			}
-			
-		}
-		
 	}
 	
-	private static class WeightsMap {
+	static class WeightsMap {
 		
-		private Map<WeightsEntry, ModelWeights> map;
+		Map<WeightsEntry, ModelWeights> map;
 		
 		public Map<String, ModelWeights> withStringKeys() {
 			return map == null ? Collections.emptyMap() : map.entrySet().stream()
 					.collect(Collectors.toMap(e -> e.getKey().toString(), e -> e.getValue()));
 		}
 
-		private static class Deserializer implements JsonDeserializer<WeightsMap> {
-
-			@Override
-			public WeightsMap deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-					throws JsonParseException {
-				
-				var weights = new WeightsMap();
-				if (json.isJsonObject()) {
-					var obj = json.getAsJsonObject();
-					var map = new LinkedHashMap<WeightsEntry, ModelWeights>();
-					for (var key : json.getAsJsonObject().keySet()) {
-						var we = WeightsEntry.fromKey(key);
-						if (we == null) {
-							logger.warn("Unsupported weights: {}", key);
-							continue;
-						}
-						map.put(we, context.deserialize(obj.get(key), ModelWeights.class));
-					}
-					weights.map = Collections.unmodifiableMap(map);
-				}
-				return weights;
-			}
-			
+		Map<WeightsEntry, ModelWeights> getMap() {
+			return map;
 		}
-		
 	}
 	
 	
@@ -687,35 +245,35 @@ public class BioimageIoSpec {
 		private URI baseURI;
 		private URI uri;
 
-		private List<InputTensor> inputs;
+		List<InputTensor> inputs;
 				
 		@SerializedName("test_inputs")
-		private List<String> testInputs;
+        List<String> testInputs;
 		@SerializedName("test_outputs")
-		private List<String> testOutputs;
+        List<String> testOutputs;
 		
 		// Should be in ISO 8601 format - but preserve string as it is found
-		private String timestamp;
+        String timestamp;
 		
-		private WeightsMap weights;
+		WeightsMap weights;
 
-		private Map<?, ?> config;
+		Map<?, ?> config;
 		
-		private List<OutputTensor> outputs;
+		List<OutputTensor> outputs;
 		
 		@SerializedName("packaged_by")
-		private List<Author> packagedBy;
+        List<Author> packagedBy;
 		
-		private ModelParent parent;
+		ModelParent parent;
 				
 		// TODO: Handle run_mode properly
 		@SerializedName("run_mode")
-		private Map<?, ?> runMode;
+        Map<?, ?> runMode;
 		
 		@SerializedName("sample_inputs")
-		private List<String> sampleInputs;
+        List<String> sampleInputs;
 		@SerializedName("sample_outputs")
-		private List<String> sampleOutputs;
+        List<String> sampleOutputs;
 
 		
 		/**
@@ -808,9 +366,7 @@ public class BioimageIoSpec {
 		}
 		
 	}
-
-
-
+	
     /**
 	 * Ensure the input is an unmodifiable list, or empty list if null.
 	 * Note that OpenJDK implementation is expected to return its input if already unmodifiable.
@@ -827,11 +383,11 @@ public class BioimageIoSpec {
 	 */
 	public static class Author {
 		
-		private String affiliation;
+		String affiliation;
 		@SerializedName("github_user")
-		private String githubUser;
-		private String name;
-		private String orcid;
+        String githubUser;
+		String name;
+		String orcid;
 		
 		public String getAffiliation() {
 			return affiliation;
@@ -859,32 +415,7 @@ public class BioimageIoSpec {
 			return base;
 		}
 		
-		
-		private static class Deserializer implements JsonDeserializer<Author> {
 
-			@Override
-			public Author deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-					throws JsonParseException {
-				
-				if (json.isJsonNull())
-					return null;
-				
-				var author = new Author();
-				if (json.isJsonObject()) {
-					var obj = json.getAsJsonObject();
-					author.name = deserializeField(context, obj, "name", String.class, null);
-					author.affiliation = deserializeField(context, obj, "affiliation", String.class, null);
-					author.githubUser = deserializeField(context, obj, "github_user", String.class, null);
-					author.orcid = deserializeField(context, obj, "orcid", String.class, null);
-				} else if (json.isJsonPrimitive()) {
-					author.name = json.getAsString();
-				}
-				return author;
-			}
-			
-		}
-		
-		
 	}
 	
 	/**
@@ -955,9 +486,12 @@ public class BioimageIoSpec {
 		
 	}
 
+	/**
+	 * The basic tensor representation for inputs and outputs.
+	 */
 	public abstract static class BaseTensor {
 
-		private Axis[] axes;
+		protected Axis[] axes;
 		@SerializedName("data_type")
 		private String dataType;
 		private String name;
@@ -1003,12 +537,57 @@ public class BioimageIoSpec {
 		}
 	}
 
+	/**
+	 * Create a shape array for a given axes.
+	 * The axes are expected to a string containing only the characters
+	 * {@code bitczyx} as defined in the spec.
+	 * <p>
+	 * The purpose of this is to build shape arrays easily without needing to
+	 * explicitly handle different axes and dimension ordering.
+	 * <p>
+	 * An example:
+	 * <pre>
+	 * <code>
+	 * int[] shape = createShapeArray("byxc", Map.of('x', 256, 'y', 512), 1);
+	 * </code>
+	 * </pre>
+	 * <p>
+	 * This should result in an int array with values {@code [1, 512, 256, 1]}.
+	 *
+	 * @param axes the axes string
+	 * @param target map defining the intended length for specified dimensions
+	 * @param defaultLength the default length to use for any dimension that are not included in the target map
+	 * @return an int array with the same length as the axes string, containing the requested dimensions or default values
+	 */
+	public static int[] createShapeArray(String axes, Map<Character, Integer> target, int defaultLength) {
+		int[] array = new int[axes.length()];
+		int i = 0;
+		for (var c : axes.toLowerCase().toCharArray()) {
+			array[i] = target.getOrDefault(c, defaultLength);
+			i++;
+		}
+		return array;
+	}
+
+
 	interface TensorDataDescription {
 
 	}
 
-	static class NominalOrOrdinalDataDescription implements TensorDataDescription {
+	/**
+	 * A description of the possible discrete data values in a tensor.
+	 */
+	public static class NominalOrOrdinalDataDescription implements TensorDataDescription {
 		private final NominalOrOrdinalDType type;
+
+		public List<? extends JsonElement> getValues() {
+			return values;
+		}
+
+		public NominalOrOrdinalDType getType() {
+			return type;
+		}
+
 		private final List<? extends JsonElement> values; // todo: can be int, float, bool, str
 
         NominalOrOrdinalDataDescription(NominalOrOrdinalDType type, List<? extends JsonElement> values) {
@@ -1017,30 +596,19 @@ public class BioimageIoSpec {
         }
 
     }
+	
 
-
-	enum NominalOrOrdinalDType {
-		FLOAT32,
-		FLOAT64,
-		UINT8,
-		INT8,
-		UINT16,
-		INT16,
-		UINT32,
-		INT32,
-		UINT64,
-		INT64,
-		BOOL
-	}
-
-	static class IntervalOrRatioDataDescription implements TensorDataDescription {
+	/**
+	 * A description of the possible ratio data values in a tensor.
+	 */
+	public static class IntervalOrRatioDataDescription implements TensorDataDescription {
 		private final IntervalOrRatioDType type;
 		private final List<Optional<Float>> range;
-		private final String unit; // todo: SI unit or "abitrary unit"
+		private final String unit; // todo: SI unit or "abitrary unit": validate somehow
 		private final float scale;
 		private final float offset;
 
-        IntervalOrRatioDataDescription(IntervalOrRatioDType type, List<Optional<Float>> range, String unit, float scale, float offset) {
+		IntervalOrRatioDataDescription(IntervalOrRatioDType type, List<Optional<Float>> range, String unit, float scale, float offset) {
             this.type = type;
             this.range = range;
             this.unit = unit;
@@ -1063,11 +631,30 @@ public class BioimageIoSpec {
 		IntervalOrRatioDataDescription(IntervalOrRatioDType type) {
 			this(type, List.of(Optional.empty(), Optional.empty()), "aritrary unit", 1, 0);
 		}
+
+		public IntervalOrRatioDType getType() {
+			return type;
+		}
+
+		public List<Optional<Float>> getRange() {
+			return range;
+		}
+
+		public String getUnit() {
+			return unit;
+		}
+
+		public float getScale() {
+			return scale;
+		}
+
+		public float getOffset() {
+			return offset;
+		}
+
 	}
-
-
-
-	enum IntervalOrRatioDType {
+	
+	public enum IntervalOrRatioDType {
 		FLOAT32,
 		FLOAT64,
 		UINT8,
@@ -1080,12 +667,26 @@ public class BioimageIoSpec {
 		INT64
 	}
 
+	public enum NominalOrOrdinalDType {
+		FLOAT32,
+		FLOAT64,
+		UINT8,
+		INT8,
+		UINT16,
+		INT16,
+		UINT32,
+		INT32,
+		UINT64,
+		INT64,
+		BOOL
+	}
+
+
 
 	/**
 	 * Model input, including shape, axes, datatype and preprocessing.
 	 */
 	public static class InputTensor extends BaseTensor {
-
 		private List<Processing> preprocessing;
 
 		public List<Processing> getPreprocessing() {
@@ -1111,7 +712,7 @@ public class BioimageIoSpec {
 			return toUnmodifiableList(postprocessing);
 		}
 
-		// todo: needs to handle halo in axis
+		// todo: needs to handle halo in axis rather than at tensor level
 		public int[] getHalo() {
 			return halo == null ? new int[0] : halo.clone();
 		}
@@ -1205,6 +806,9 @@ public class BioimageIoSpec {
 
 		private void validate(List<? extends BaseTensor> otherTensors) {}
 
+		/**
+		 * A shape that is determined based on a minimum and a step size.
+		 */
 		public static class ParameterizedInputShape extends Shape {
 			
 			private int[] min;
@@ -1251,7 +855,10 @@ public class BioimageIoSpec {
 			}
 
 		}
-		
+
+		/**
+		 * A shape that is determined based on the shape of another tensor.
+		 */
 		public static class ImplicitOutputShape extends Shape {
 			
 			private String referenceTensor;
@@ -1382,14 +989,6 @@ public class BioimageIoSpec {
 	}
 
 
-	/**
-	 * Get the old "bcyx" style axis representation of an Axis array.
-	 * @param axes The Axis array.
-	 * @return A string representing the axis types.
-	 */
-	public static String getAxesString(Axis[] axes) {
-		return Arrays.stream(axes).map(a -> a.getType().toString()).collect(Collectors.joining());
-	}
 	
 	/**
 	 * Base class for pre- and post-processing operations.
@@ -1400,10 +999,10 @@ public class BioimageIoSpec {
 		/**
 		 * Requested mode for pre- or post-processing.
 		 */
-		public enum ProcessingMode { FIXED, PER_DATASET, PER_SAMPLE }
+		public enum ProcessingMode {FIXED, PER_DATASET, PER_SAMPLE}
 
 		private final String name;
-		private Map<String, Object> kwargs;
+		Map<String, Object> kwargs;
 
 		Processing(String name) {
 			this.name = name;
@@ -1417,14 +1016,16 @@ public class BioimageIoSpec {
 			return kwargs == null ? Collections.emptyMap() : Collections.unmodifiableMap(kwargs);
 		}
 
-
+		/**
+		 * A binarize operation.
+		 */
 		public static class Binarize extends Processing {
 
 			Binarize() {
 				super("binarize");
 			}
 
-			private double threshold = Double.NaN;
+			double threshold = Double.NaN;
 
 			public double getThreshold() {
 				return threshold;
@@ -1432,10 +1033,13 @@ public class BioimageIoSpec {
 
 		}
 
+		/**
+		 * A clipping operation.
+		 */
 		public static class Clip extends Processing {
 
-			private double min;
-			private double max;
+			double min;
+			double max;
 
 			Clip() {
 				super("clip");
@@ -1451,11 +1055,14 @@ public class BioimageIoSpec {
 
 		}
 
+		/**
+		 * A linear scaling operation.
+		 */
 		public static class ScaleLinear extends Processing {
 
-			private double[] gain;
-			private double[] offset;
-			private Axis[] axes;
+			double[] gain;
+			double[] offset;
+			Axis[] axes;
 
 			ScaleLinear() {
 				super("scale_linear");
@@ -1475,6 +1082,9 @@ public class BioimageIoSpec {
 
 		}
 
+		/**
+		 * A sigmoid transofmration.
+		 */
 		public static class Sigmoid extends Processing {
 
 			Sigmoid() {
@@ -1485,9 +1095,9 @@ public class BioimageIoSpec {
 
 		protected abstract static class ProcessingWithMode extends Processing {
 
-			private ProcessingMode mode = ProcessingMode.PER_SAMPLE;
-			private Axis[] axes;
-			private double eps = 1e-6;
+			ProcessingMode mode = ProcessingMode.PER_SAMPLE;
+			Axis[] axes;
+			double eps = 1e-6;
 
 			ProcessingWithMode(String name) {
 				super(name);
@@ -1507,10 +1117,12 @@ public class BioimageIoSpec {
 
 		}
 
-
+		/**
+		 * A scaling operation to the mean and variance of a reference tensor.
+		 */
 		public static class ScaleMeanVariance extends ProcessingWithMode {
 
-			private String referenceTensor;
+			String referenceTensor;
 
 			ScaleMeanVariance() {
 				super("scale_mean_variance");
@@ -1522,13 +1134,15 @@ public class BioimageIoSpec {
 
 		}
 
-
+		/**
+		 * A scaling operation to a reference tensor.
+		 */
 		public static class ScaleRange extends ProcessingWithMode {
 
-			private String referenceTensor; // TODO: Figure out whether to use this somehow
+			String referenceTensor; // TODO: Figure out whether to use this somehow
 
-			private double minPercentile = 0.0;
-			private double maxPercentile = 100.0;
+			double minPercentile = 0.0;
+			double maxPercentile = 100.0;
 
 			ScaleRange() {
 				super("scale_range");
@@ -1549,11 +1163,13 @@ public class BioimageIoSpec {
 
 		}
 
-
+		/**
+		 * A scaling operation to zero mean and unit variqnce.
+		 */
 		public static class ZeroMeanUnitVariance extends ProcessingWithMode {
 
-			private double[] mean;
-			private double[] std;
+			double[] mean;
+			double[] std;
 
 			ZeroMeanUnitVariance() {
 				super("zero_mean_unit_variance");
@@ -1569,197 +1185,15 @@ public class BioimageIoSpec {
 
 		}
 
-		private static class ProcessingDeserializer implements JsonDeserializer<Processing> {
-
-			@Override
-			public Processing deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-					throws JsonParseException {
-				
-				if (json.isJsonNull())
-					return null;
-				
-				var obj = json.getAsJsonObject();
-				var name = obj.has("name") ? obj.get("name").getAsString() : obj.get("id").getAsString();
-				var kwargs = obj.has("kwargs") ? obj.get("kwargs").getAsJsonObject() : null;
-				switch (name) {
-				case "binarize":
-					var binarize = new Binarize();
-					binarize.threshold = deserializeField(context, kwargs, "threshold", Double.class, true);
-					return binarize;
-				case "clip":
-					var clip = new Clip();
-					clip.min = deserializeField(context, kwargs, "min", Double.class, Double.NEGATIVE_INFINITY);
-					clip.max = deserializeField(context, kwargs, "max", Double.class, Double.POSITIVE_INFINITY);
-					return clip;
-				case "scale_linear":
-					var scaleLinear = new ScaleLinear();
-					scaleLinear.axes = deserializeField(context, kwargs, "axes", Axis[].class, false);
-					scaleLinear.gain = deserializeField(context, kwargs, "gain", double[].class, false);
-					scaleLinear.offset = deserializeField(context, kwargs, "offset", double[].class, false);
-					return scaleLinear;
-				case "scale_mean_variance":
-					var scaleMeanVariance = new ScaleMeanVariance();
-					((ProcessingWithMode)scaleMeanVariance).mode = deserializeField(context, kwargs, "mode", ProcessingMode.class, false);
-					((ProcessingWithMode)scaleMeanVariance).axes = deserializeField(context, kwargs, "axes", Axis[].class, false);
-					((ProcessingWithMode)scaleMeanVariance).eps = deserializeField(context, kwargs, "eps", Double.class, 1e-6);
-					scaleMeanVariance.referenceTensor = deserializeField(context, kwargs, "reference_tensor", String.class, null);
-					return scaleMeanVariance;
-				case "scale_range":
-					var scaleRange = new ScaleRange();
-					((ProcessingWithMode)scaleRange).mode = deserializeField(context, kwargs, "mode", ProcessingMode.class, false);
-					((ProcessingWithMode)scaleRange).axes = deserializeField(context, kwargs, "axes", Axis[].class, false);
-					((ProcessingWithMode)scaleRange).eps = deserializeField(context, kwargs, "eps", Double.class, 1e-6);
-					scaleRange.referenceTensor = deserializeField(context, obj, "reference_tensor", String.class, null);
-					scaleRange.maxPercentile = deserializeField(context, kwargs, "max_percentile", Double.class, 0.0);
-					scaleRange.minPercentile = deserializeField(context, kwargs, "min_percentile", Double.class, 100.0);
-					return scaleRange;
-				case "sigmoid":
-					return new Sigmoid();
-				case "zero_mean_unit_variance":
-					var zeroMeanUnitVariance = new ZeroMeanUnitVariance();
-					((ProcessingWithMode)zeroMeanUnitVariance).mode = deserializeField(context, kwargs, "mode", ProcessingMode.class, false);
-					((ProcessingWithMode)zeroMeanUnitVariance).axes = deserializeField(context, kwargs, "axes", Axis[].class, false);
-					((ProcessingWithMode)zeroMeanUnitVariance).eps = deserializeField(context, kwargs, "eps", Double.class, 1e-6);
-					zeroMeanUnitVariance.mean = deserializeField(context, kwargs, "mean", double[].class, false);
-					zeroMeanUnitVariance.std = deserializeField(context, kwargs, "std", double[].class, false);
-					return zeroMeanUnitVariance;
-				default:
-					var processing = new Processing(name);
-					processing.kwargs = kwargs == null ? Collections.emptyMap() : context.deserialize(kwargs, Map.class);
-					return processing;
-				}
-			}
-			
-		}
-		
-		
-		private static class ModeDeserializer implements JsonDeserializer<ProcessingMode> {
-
-			@Override
-			public ProcessingMode deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-					throws JsonParseException {
-				
-				if (json.isJsonNull())
-					return null;
-				
-				var s = json.getAsString();
-				for (var mode : ProcessingMode.values()) {
-					if (s.equals(mode.name()))
-						return mode;
-				}
-
-				for (var mode : ProcessingMode.values()) {
-					if (s.equalsIgnoreCase(mode.name()))
-						return mode;
-				}
-				
-				logger.warn("Unknown processing mode: {}", s);
-				return null;
-			}
-			
-		}
-		
 	}
 
-	private static Size deserializeSize(JsonDeserializationContext context, JsonElement jsonElement, JsonElement scale) {
-		if (jsonElement == null || jsonElement.isJsonNull()) {
-			return null;
-		}
-		if (jsonElement.isJsonPrimitive()) {
-			return new FixedSize(jsonElement.getAsInt());
-		}
-		if (jsonElement.isJsonObject()) {
-			JsonObject obj = jsonElement.getAsJsonObject();
-			if (obj.has("min") && obj.has("step")) {
-				return context.deserialize(obj, ParameterizedSize.class);
-			}
-			if (obj.has("axis_id") && obj.has("tensor_id")) {
-				return new ReferencedSize(
-						obj.get("axis_id").getAsString(),
-						obj.get("tensor_id").getAsString(),
-						scale != null ? scale.getAsDouble() : 1.0,
-						deserializeField(context, obj, "offset", Integer.class, 1)
-				);
-			}
-			return new DataDependentSize(
-					deserializeField(context, obj, "min", Integer.class, 1),
-					deserializeField(context, obj, "max", Integer.class, 1));
-		}
-        logger.error("Unknown JSON element {}", jsonElement);
-		throw new JsonParseException("No idea what type of size this is, sorry!");
-	}
-
-	public static class AxesDeserializer implements JsonDeserializer<Axis[]> {
-
-		@Override
-		public Axis[] deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext context) throws JsonParseException {
-			if (jsonElement.isJsonPrimitive()) {
-				var s = jsonElement.getAsString();
-				Axis[] axes = new Axis[s.length()];
-				for (int i = 0; i < axes.length; i++) {
-					axes[i] = new CharAxis(s.charAt(i));
-				}
-				return axes;
-			}
-			// todo: input or output???
-			if (jsonElement.isJsonArray()) {
-				var arr = jsonElement.getAsJsonArray();
-				Axis[] axes = new Axis[arr.size()];
-				for (int i = 0; i < axes.length; i++) {
-					var curr = arr.get(i);
-					if (curr.isJsonPrimitive()) {
-						axes[i] = new CharAxis(curr.getAsString().charAt(0));
-						continue;
-					}
-
-					var oj = curr.getAsJsonObject();
-					var id = deserializeField(context, oj, "id", String.class, "");
-					var desc = deserializeField(context, oj, "description", String.class, "");
-					Size size = deserializeSize(context, oj.get("size"), oj.get("scale"));
-					switch (oj.get("type").getAsString()) {
-						case "time":
-							axes[i] = new TimeInputAxis(
-									id, desc,
-									TimeUnit.valueOf(oj.get("unit").getAsString().toUpperCase()),
-									oj.get("scale").getAsDouble(),
-									size
-							);
-							break;
-						case "channel":
-							var namesJSON = oj.get("channel_names").getAsJsonArray();
-							List<String> names = new LinkedList<>();
-							for (JsonElement n : namesJSON) {
-								names.add(n.getAsString());
-							}
-							axes[i] = new ChannelAxis(
-									id, desc,
-									names
-							);
-							break;
-						case "index":
-							axes[i] = new IndexInputAxis(id, desc, size);
-							break;
-						case "space":
-							axes[i] = new SpaceInputAxis(
-									id, desc,
-									deserializeField(context, oj, "unit", String.class, ""),
-									deserializeField(context, oj, "scale", Double.class, 1),
-									size
-							);
-							break;
-						case "batch":
-							axes[i] = new BatchAxis(id, desc, deserializeField(context, oj, "size", Integer.class, 1));
-							break;
-						default:
-							logger.error("Unknown object {}", oj);
-							axes[i] = null;
-					}
-				}
-				return axes;
-			}
-			logger.error("Unknown JSON element {}", jsonElement);
-			return null;
-		}
+	/**
+	 * Get the old "bcyx" style axis representation of an Axis array.
+	 * @param axes The Axis array.
+	 * @return A string representing the axis types.
+	 */
+	public static String getAxesString(Axis[] axes) {
+		return Arrays.stream(axes).map(a -> a.getType().toString()).collect(Collectors.joining());
 	}
 
 	/**
@@ -1816,6 +1250,66 @@ public class BioimageIoSpec {
 		public String toString() {
 			return type;
 		}
+	}
+
+
+	/**
+	 * Axes that have both size and scale.
+	 */
+	public interface ScaledAxis {
+		/**
+		 * Get the size of this axis.
+		 * @return The size.
+		 */
+		Size getSize();
+
+		/**
+		 * Get the scale of this axis.
+		 * @return The scale (might be constant).
+		 */
+		double getScale();
+	}
+
+	/**
+	 * An axis size. Can be fixed, data-dependent (unknown), parameterized, or based on another axis.
+	 */
+	public interface Size {
+		int NO_SIZE = -1;
+
+		/**
+		 * Get the default size of this axis. {@link #getTargetSize(int)} may be more useful.
+		 * @return The size of this axis.
+		 */
+		int getSize();
+
+		/**
+		 * Get a size as close as possible to a target.
+		 * @param target The target size.
+		 * @return The fixed output size, {@link #NO_SIZE} or as close as we can get to the target size.
+		 */
+		int getTargetSize(int target);
+
+		/**
+		 * Get the step size of this axis size.
+		 * @return {@link #NO_SIZE} for any axis without a step size, otherwise the size.
+		 */
+		int getStep();
+
+		void validate(List<? extends BaseTensor> tensors);
+	}
+
+	/**
+	 * An axis with a halo.
+	 * The halo should be cropped from the output tensor to avoid boundary effects.
+	 * It is to be cropped from both sides, i.e. `size_after_crop = size - 2 * halo`.
+	 * To document a halo that is already cropped by the model use `size.offset` instead.
+	 */
+	public interface WithHalo {
+		/**
+		 * Get the size of the halo for this axis.
+		 * @return The size of the halo in pixels.
+		 */
+		int getHalo();
 	}
 
 	/**
@@ -1973,50 +1467,6 @@ public class BioimageIoSpec {
 		}
 	}
 
-	/**
-	 * Axes that have both size and scale.
-	 */
-	public interface ScaledAxis {
-		/**
-		 * Get the size of this axis.
-		 * @return The size.
-		 */
-		Size getSize();
-
-		/**
-		 * Get the scale of this axis.
-		 * @return The scale (might be constant).
-		 */
-		double getScale();
-	}
-
-	/**
-	 * An axis size. Can be fixed, data-dependent (unknown), parameterized, or based on another axis.
-	 */
-	public interface Size {
-		int NO_SIZE = -1;
-
-		/**
-		 * Get the default size of this axis. {@link #getTargetSize(int)} may be more useful.
-		 * @return The size of this axis.
-		 */
-		int getSize();
-
-		/**
-		 * Get a size as close as possible to a target.
-		 * @param target The target size.
-		 * @return The fixed output size, {@link #NO_SIZE} or as close as we can get to the target size.
-		 */
-		int getTargetSize(int target);
-
-		/**
-		 * Get the step size of this axis size.
-		 * @return {@link #NO_SIZE} for any axis without a step size, otherwise the size.
-		 */
-		int getStep();
-
-		void validate(List<? extends BaseTensor> tensors);
-	}
 
 	/**
 	 * Describes a range of valid tensor axis sizes as `size = min + n*step`.
@@ -2193,9 +1643,12 @@ public class BioimageIoSpec {
 			return scale;
 		}
 
+		public TimeUnit getUnit() {
+			return unit;
+		}
 	}
 
-	public static class TimeInputAxis extends TimeAxisBase {
+	static class TimeInputAxis extends TimeAxisBase {
 		private final Size size;
 
 		TimeInputAxis(String id, String description, TimeUnit unit, double scale, Size size) {
@@ -2214,7 +1667,7 @@ public class BioimageIoSpec {
 		}
 	}
 
-	public static class TimeOutputAxis extends TimeAxisBase {
+	static class TimeOutputAxis extends TimeAxisBase {
 		private final Size size;
 
 		TimeOutputAxis(String id, String description, TimeUnit unit, double scale, Size size) {
@@ -2226,13 +1679,14 @@ public class BioimageIoSpec {
 		public Size getSize() {
 			return size;
 		}
+
 		@Override
 		public void validate(List<? extends BaseTensor> tensors) {
 			getSize().validate(tensors);
 		}
 	}
 
-	public static class TimeOutputAxisWithHalo extends TimeOutputAxis implements WithHalo {
+	static class TimeOutputAxisWithHalo extends TimeOutputAxis implements WithHalo {
 		private final int halo;
 
 		TimeOutputAxisWithHalo(String id, String description, TimeUnit unit, double scale, Size size, int halo) {
@@ -2308,17 +1762,7 @@ public class BioimageIoSpec {
 		}
 	}
 
-	public enum BioImageIoVersion {
-		VERSION_0_4("0.4"),
-		VERSION_0_5("0.5");
-		private final String version;
-
-		BioImageIoVersion(String s) {
-			this.version = s;
-		}
-	}
-
-	public abstract static class SpaceAxisBase extends AxisBase implements ScaledAxis {
+	abstract static class SpaceAxisBase extends AxisBase implements ScaledAxis {
 		private final SpaceUnit unit;
 		private double scale = 1.0;
 
@@ -2341,9 +1785,13 @@ public class BioimageIoSpec {
 		public double getScale() {
 			return this.scale;
 		}
-	}
 
-	public static class SpaceInputAxis extends SpaceAxisBase {
+        public SpaceUnit getUnit() {
+            return unit;
+        }
+    }
+
+	static class SpaceInputAxis extends SpaceAxisBase {
 		private final Size size;
 		private final boolean concatenable = false;
 
@@ -2362,7 +1810,7 @@ public class BioimageIoSpec {
 		}
 	}
 
-	public static class SpaceOutputAxis extends SpaceAxisBase {
+	static class SpaceOutputAxis extends SpaceAxisBase {
 		private final Size size;
 
 		SpaceOutputAxis(String id, String description, String unit, double scale, Size size) {
@@ -2390,105 +1838,24 @@ public class BioimageIoSpec {
 			this.halo = halo;
 		}
 
-		public int getHalo() {
+		@Override
+        public int getHalo() {
 			return this.halo;
 		}
 	}
 
-	/**
-	 * An axis with a halo.
-	 * The halo should be cropped from the output tensor to avoid boundary effects.
-	 * It is to be cropped from both sides, i.e. `size_after_crop = size - 2 * halo`.
-	 * To document a halo that is already cropped by the model use `size.offset` instead.
-	 */
-	interface WithHalo {
-		/**
-		 * Get the size of the halo for this axis.
-		 * @return The size of the halo in pixels.
-		 */
-		int getHalo();
-	}
 
-	/**
-	 * Deal with the awkwardness of -inf and inf instead of .inf and .inf.
-	 * This otherwise caused several failures for data range.
-	 */
-	private static class DoubleArrayDeserializer implements JsonDeserializer<double[]> {
+	public enum BioImageIoVersion {
+		VERSION_0_4("0.4"),
+		VERSION_0_5("0.5");
+		private final String version;
 
-		@Override
-		public double[] deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-				throws JsonParseException {
+		BioImageIoVersion(String s) {
+			this.version = s;
+		}
 
-			if (json.isJsonNull())
-				return null;
-			if (json.isJsonArray()) {
-				List<Double> values = new ArrayList<>();
-				for (var jsonVal : json.getAsJsonArray()) {
-					if (jsonVal.isJsonNull()) {
-						logger.warn("Found null when expecting a double - will replace with NaN");
-						values.add(Double.NaN);
-						continue;
-					}
-					var jsonPrimitive = jsonVal.getAsJsonPrimitive();
-					if (jsonPrimitive.isNumber()) {
-						values.add(jsonPrimitive.getAsDouble());
-					} else {
-						var s = jsonPrimitive.getAsString();
-						if ("inf".equalsIgnoreCase(s))
-							values.add(Double.POSITIVE_INFINITY);
-						else if ("-inf".equalsIgnoreCase(s))
-							values.add(Double.NEGATIVE_INFINITY);
-						else
-							values.add(Double.parseDouble(s));
-					}
-				}
-				return values.stream().mapToDouble(d -> d).toArray();
-			} else
-				throw new JsonParseException("Can't parse data range from " + json);
+		public String getVersion() {
+			return version;
 		}
 	}
-
-	private static class TensorDataDescriptionDeserializer implements JsonDeserializer<TensorDataDescription> {
-
-		@Override
-		public TensorDataDescription deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext context) throws JsonParseException {
-			if (jsonElement.isJsonNull()) {
-				return null;
-			}
-			if (jsonElement.isJsonObject()) {
-				JsonObject jsonObject = jsonElement.getAsJsonObject();
-				if (jsonObject.get("values") != null) {
-					JsonArray vals = jsonObject.get("values").getAsJsonArray(); // todo: parse to values...
-					return new NominalOrOrdinalDataDescription(
-							NominalOrOrdinalDType.valueOf(jsonObject.get("type").getAsString().toUpperCase()),
-							vals.asList()
-					);
-				}
-				var t = jsonObject.get("type");
-				var r = jsonObject.get("range");
-				List<Optional<Float>> range;
-				if (r == null) {
-					range = List.of(Optional.empty(), Optional.empty());
-				} else {
-					range = r.getAsJsonArray().asList().stream()
-							.map(JsonElement::getAsFloat)
-							.map(Optional::of)
-							.collect(Collectors.toList());
-				}
-				JsonElement unit = jsonObject.get("unit");
-				JsonElement scale = jsonObject.get("scale");
-				JsonElement offset = jsonObject.get("offset");
-				return new IntervalOrRatioDataDescription(
-                        IntervalOrRatioDType.valueOf((t != null ? t.getAsString(): "float32").toUpperCase()),
-						range,
-						unit != null ? unit.getAsString() : "abitrary unit",
-						scale != null ? scale.getAsFloat() :  1.0f,
-						offset != null ? offset.getAsFloat() :  1.0f
-				);
-			}
-			logger.warn("Unknown data description! Returning null");
-			return null;
-		}
-	}
-
 }
